+++ src/lib.rs
@@ -4,7 +4,10 @@
 
 pub mod component_v0_4;
 pub mod host_import_v0_2;
+pub mod host_import_v0_4;
 pub mod pack_export_v0_2;
+pub mod pack_export_v0_4;
 pub mod secrets_v0_1;
 pub mod types_core_v0_2;
+pub mod types_core_v0_4;
 pub mod wasix_mcp_v0_0_5;